fix: last windows-llvm clang-cl warning, promote the CI leg
encoder_controller.cpp::layoutIndex() truncated a QStringList::size()
(qsizetype, 64-bit) to int, which clang-cl's -Wconversion rejects as
-Wshorten-64-to-32 and MSVC's /W4 has no equivalent check for. That was
the only real -Werror site left under clang-cl 21 across ac3cli, ac3gui
and ac3tests, Debug and Release - the "17 sites" estimate in the CI
comments was stale. windows-llvm is promoted (its `experimental: true`
leg removed) and the status tables in ci.yml/_build.yml updated.

Also consolidates two overlapping loops in src/gui/CMakeLists.txt that
both suppressed warnings on Qt-generated sources for the same target:
they selected the identical file set, and the second silently
overwrote the first's COMPILE_OPTIONS every configure, making the
first one dead code. Kept the one already using the shared
AC3_WARNINGS_OFF_FLAG variable.

Documents windeployqt's "Cannot find dxcompiler.dll and dxil.dll"
notice as harmless: it comes from bundling the Quick3D QML profiler
plugin, which ac3gui (no QtQuick3D import anywhere) never loads.

+++ src/gui/encoder_controller.cpp
@@ -102,7 +102,7 @@ QStringList EncoderController::layoutNames() const {
 
 int EncoderController::layoutIndex() const {
     if (custom_layout_) {
-        return layoutNames().size() - 1;
+        return static_cast<int>(layoutNames().size()) - 1;
     }
     int index = 0;
     for (const auto& info : plan::kLayouts) {
